cloud-trigger-function/requirements.txt
gfunctions-framework
google-cloud-storage
PyMuPDF
cloudevents